backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.2",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "hyperbrowser>=0.45.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.63",
    "langchain-openai>=0.3.19",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
